+++ MilkTeaHouseProject/DTO/Staff.cs
@@ -16,7 +16,7 @@ namespace MilkTeaHouseProject.DTO
         private string position;
         private string userName;
         private int overTime;
-        private int salaryReceived;
+        private long salaryReceived;
         private int fault;
         private string phoneNumber;
         private string cmnd;
@@ -30,14 +30,14 @@ namespace MilkTeaHouseProject.DTO
         public string UserName { get => userName; set => userName = value; }
         public int OverTime { get => overTime; set => overTime = value; }
         public byte[] Image { get => image; set => image = value; }
-        public int SalaryReceived { get => salaryReceived; set => salaryReceived = value; }
+        public long SalaryReceived { get => salaryReceived; set => salaryReceived = value; }
         public int Fault { get => fault; set => fault = value; }
         public string PhoneNumber { get => phoneNumber; set => phoneNumber = value; }
         public string CMND { get => cmnd; set => cmnd = value; }
         public string Address { get => address; set => address = value; }
         public bool Sex { get => sex; set => sex = value; }
 
-        public Staff(int id, string name, byte[] image, DateTime birthdate, string position, string username, int overtime, int salaryreceived, int fault, string phonenumber, string cmnd, string address, bool sex)
+        public Staff(int id, string name, byte[] image, DateTime birthdate, string position, string username, int overtime, long salaryreceived, int fault, string phonenumber, string cmnd, string address, bool sex)
         {
             this.iD = id;
             this.name = name;
@@ -69,7 +69,7 @@ namespace MilkTeaHouseProject.DTO
             this.cmnd = row["CMND"].ToString();
             this.address = row["Address"].ToString();
             this.sex = (bool)row["Sex"];
-            this.salaryReceived = int.Parse(row["SalaryReceived"].ToString());
+            this.salaryReceived = long.Parse(row["SalaryReceived"].ToString());
         }
     }
 }
